requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
colorama==0.4.6
pycryptodome==3.22.0
dataclass-factory==2.16
decorator==5.2.1
future==1.0.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
httplib2==0.22.0
imageio==2.37.0
langdetect==1.0.9
mega.py==1.0.8
numpy==2.2.2
pandas==2.2.3
pg8000==1.31.2
Pillow==11.1.0
pydantic==2.9.2
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
tenacity==5.1.5
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.29.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0.1
websocket-client==1.8.0